requirements.txt
aerich==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
aiohttp==3.8.3 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.10" and python_version < "4.0"
beautifulsoup4==4.11.2 ; python_version >= "3.10" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==39.0.1 ; python_version >= "3.10" and python_version < "4.0"
dapr==1.8.3 ; python_version >= "3.10" and python_version < "4.0"
dictdiffer==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi-auth0==0.3.2 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.90.0 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.48.2 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
loader==2017.9.11 ; python_version >= "3.10" and python_version < "4.0"
lxml==4.9.2 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
nest-asyncio==1.5.6 ; python_version >= "3.10" and python_version < "4.0"
pdfminer-six==20221105 ; python_version >= "3.10" and python_version < "4.0"
pdfplumber==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.4 ; python_version >= "3.10" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-docx==0.8.11 ; python_version >= "3.10" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
soupsieve==2.4 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.23.0 ; python_version >= "3.10" and python_version < "4.0"
tomlkit==0.11.6 ; python_version >= "3.10" and python_version < "4.0"
tortoise-orm==0.19.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
wand==0.6.11 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
chardet==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi_pagination==0.11.4 ; python_version >= "3.10" and python_version < "4.0"
pycryptodome==3.17 ; python_version >= "3.10" and python_version < "4.0"
mixpanel==4.10.0;python_version >= "3.10" and python_version < "4.0"
fastapi~=0.90.0
starlette~=0.23.0
pydantic~=1.10.4
auth0-python~=4.0.0